+++ polaris-agent-adapter/pinpoint/pinpoint-nacos-plugin/src/main/java/cn/polarismesh/agent/adapter/nacos/interceptor/NacosNamingFactoryInterceptor.java
@@ -1,0 +1,11 @@
+package cn.polarismesh.agent.adapter.nacos.interceptor;
+
+
+/**
+ * nacos sdk1.x版本双注册拦截类
+ *
+ * @author bruceppeng
+ */
+public class NacosNamingFactoryInterceptor extends AbstractNacosInterceptor {
+
+}
