+++ polaris-agent-adapter/pinpoint/pinpoint-nacos-plugin/src/main/resources/META-INF/services/com.navercorp.pinpoint.bootstrap.plugin.ProfilerPlugin
@@ -1,0 +1,1 @@
+cn.polarismesh.agent.adapter.nacos.NacosDoubleRegistryPlugin
