--- polaris-agent-adapter/pinpoint/pinpoint-nacosv1-plugin/pom.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>polaris-agent-adapter-pinpoint</artifactId>
-        <groupId>com.tencent.polaris</groupId>
-        <version>${revision}</version>
-        <relativePath>../pom.xml</relativePath>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>pinpoint-polaris-nacosv1-plugin</artifactId>
-
-    <name>Pinpoint Nacosv1 Plugin</name>
-    <dependencies>
-        <!-- nacos-double-registry-plugins-->
-        <dependency>
-            <groupId>com.tencent.polaris</groupId>
-            <artifactId>polaris-agent-nacosv1-plugin</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.tencent.polaris</groupId>
-            <artifactId>pinpoint-polaris-common</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-    </dependencies>
-</project>
