--- polaris-agent-adapter/pinpoint/pinpoint-nacosv1-plugin/src/main/java/cn/polarismesh/agent/adapter/nacos/v1/interceptor/InterceptorBuilder.java
@@ -1,36 +1,0 @@
-/*
- * Tencent is pleased to support the open source community by making Polaris available.
- *
- * Copyright (C) 2019 THL A29 Limited, a Tencent company. All rights reserved.
- *
- * Licensed under the BSD 3-Clause License (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://opensource.org/licenses/BSD-3-Clause
- *
- * Unless required by applicable law or agreed to in writing, software distributed
- * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
- * CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package cn.polarismesh.agent.adapter.nacos.v1.interceptor;
-
-
-import cn.polarismesh.agent.core.nacos.v1.interceptor.NacosV1NamingFactoryInterceptor;
-import cn.polarismesh.pinpoint.common.InterceptorFactory;
-
-/**
- * nacos interceptor builder
- *
- * @author bruceppeng
- */
-public class InterceptorBuilder {
-
-    static void buildInterceptors() {
-        NacosV1NamingFactoryInterceptor nacosV1NamingFactoryInterceptor = new NacosV1NamingFactoryInterceptor();
-
-        InterceptorFactory.addInterceptor(NacosNamingFactoryInterceptor.class, nacosV1NamingFactoryInterceptor);
-    }
-}
