--- polaris-agent-adapter/pinpoint/pinpoint-nacosv1-plugin/src/main/java/cn/polarismesh/agent/adapter/nacos/v1/interceptor/NacosNamingFactoryInterceptor.java
@@ -1,11 +1,0 @@
-package cn.polarismesh.agent.adapter.nacos.v1.interceptor;
-
-
-/**
- * nacos sdk1.x版本双注册拦截类
- *
- * @author bruceppeng
- */
-public class NacosNamingFactoryInterceptor extends AbstractNacosInterceptor {
-
-}
