--- polaris-agent-adapter/pinpoint/pinpoint-nacosv1-plugin/src/main/resources/META-INF/services/com.navercorp.pinpoint.bootstrap.plugin.ProfilerPlugin
@@ -1,1 +1,0 @@
-cn.polarismesh.agent.adapter.nacos.v1.NacosDoubleRegistryPlugin
