+++ polaris-agent-adapter/pinpoint/pom.xml
@@ -14,7 +14,7 @@
     <name>Polaris Agent Adapter Pinpoint</name>
     <modules>
         <!--<module>pinpoint-spring-cloud-2021-plugin</module>-->
-        <module>pinpoint-nacosv1-plugin</module>
+        <module>pinpoint-nacos-plugin</module>
         <module>pinpoint-dubbo2-plugin</module>
         <module>pinpoint-dubbox-plugin</module>
         <module>pinpoint-bootstrap</module>
