+++ polaris-agent-build/bin/plugins
@@ -1,1 +1,1 @@
-nacosv1
+nacos
