+++ polaris-agent-core/plugins/nacos/pom.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>polaris-agent-core-plugins</artifactId>
+        <groupId>com.tencent.polaris</groupId>
+        <version>${revision}</version>
+        <relativePath>../pom.xml</relativePath>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>polaris-agent-nacos-plugin</artifactId>
+
+    <name>Polaris Agent Nacos Plugin</name>
+
+    <properties>
+        <spring.cloud.version>2021.0.0</spring.cloud.version>
+        <spring.vesion>5.3.13</spring.vesion>
+        <spring.boot.vesion>2.6.0</spring.boot.vesion>
+        <ribbon.vesion>2.2.10.RELEASE</ribbon.vesion>
+        <slf4j.vesion>1.7.30</slf4j.vesion>
+    </properties>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.alibaba.nacos</groupId>
+            <artifactId>nacos-client</artifactId>
+            <version>${nacos.client.version}</version>
+        </dependency>
+    </dependencies>
+</project>
