--- polaris-agent-core/plugins/nacosv1/pom.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>polaris-agent-core-plugins</artifactId>
-        <groupId>com.tencent.polaris</groupId>
-        <version>${revision}</version>
-        <relativePath>../pom.xml</relativePath>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>polaris-agent-nacosv1-plugin</artifactId>
-
-    <name>Polaris Agent Nacosv1 Plugin</name>
-
-    <properties>
-        <spring.cloud.version>2021.0.0</spring.cloud.version>
-        <spring.vesion>5.3.13</spring.vesion>
-        <spring.boot.vesion>2.6.0</spring.boot.vesion>
-        <ribbon.vesion>2.2.10.RELEASE</ribbon.vesion>
-        <slf4j.vesion>1.7.30</slf4j.vesion>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.alibaba.nacos</groupId>
-            <artifactId>nacos-client</artifactId>
-            <version>${nacos.v1.client.version}</version>
-        </dependency>
-    </dependencies>
-</project>
